changes in logic and styles
