+++ src/components/organisms/TravelersForm.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React from 'react';
 import { useFormContext, Controller } from 'react-hook-form';
 import TravelerFields from '../molecules/TravelerFields';
