+++ src/components/templates/WizardStepTemplate.tsx
@@ -7,7 +7,7 @@ interface WizardStepTemplateProps {
 }
 
 const WizardStepTemplate: React.FC<WizardStepTemplateProps> = ({ title, description, children }) => (
-  <section className="min-h-screen flex flex-col items-center justify-center bg-[#03230d] py-8 px-4">
+  <section className="min-h-screen flex flex-col items-center justify-center py-8 px-4">
     <div className="w-full max-w-2xl mx-auto">
       <div className="bg-[#015c30] rounded-2xl shadow-lg p-8 min-h-[75vh]">
         <h1 className="text-3xl font-bold mb-2 text-white text-center drop-shadow-lg">{title}</h1>
